Update main.js
for loop to iterate through image file and insert an img element inside the thumb bar class that embeds that image in the page along with its alternative text. Added a click event listener to each img inside the thumb-bar so that, when they are clicked, the corresponding image and alternative text are displayed in the displayed img element.