Update main.js
Created update method to actually move the ball. The first two if statement to check if the ball position of the ball is not above the  left or right edges of the canvas. .While the last 2 if statement to check if y position of the ball does not pass the top or bottom edges of the canvas.. if yes the position of the ball is reversed. this.x += this.velX;
        this.y += this.velY; to update the velocity the ball moves