Update main-finished.js
A new class called EvilCircle created which also inherits from shape class.  passed x and y in my constructor and passed the x, y arguments up to the Shape superclass along with values for velX and velY hardcoded to 20. set color to white and size to 10.  keydown eventlistener created in a way that if the the specified keys are pressed the evil circle moves to right or left or up or down.