Update main-finished.js
checkbound() method created to update the location of the evil circle . the first two if statement to update x axes location and the second if statement update y axes location so that it does not go off-screen. collisiondetect() method also created in a way that when ball collide with the evil circle it stop existing.